requirements.txt
python-telegram-bot
python-dotenv
pytesseract
pdf2image
Pillow